backend/requirements.txt
Django==4.2
djangorestframework
django-environ
requests
django-log-request-id
django_extensions
django-allauth
django-cors-headers
psycopg2-binary
python-jose
cryptography
pycryptodome
gunicorn
